test: Bail out of test if API returns status code 400 after retries
Testing has shown that retrying the upload (of a valid file) did not
keep the action from failing.
This behavior could not be reproduced locally and only happens in CI
from time to time.
As we know the tests work, we bail out after several retries and do not
fail the test if it still returns status code 400 (something is not
"right" with the uploaded file).

Signed-off-by: David Runge <dvzrv@archlinux.org>