fix: Simplify use of strum macros for SignatureType
Rely on a simpler approach for `FromStr` implementations provided via
strum macros by using "ASCII, case insensitive" for all members of the
`SignatureType` enum.
This approach is in line with all other custom enum types.

Add documentation for all members of `SignatureType`.
Update documentation of `SignatureType` to correctly reflect the ECDSA
type members of the enum.

Signed-off-by: David Runge <dvzrv@archlinux.org>