feat!: Assemble connection configuration only when it is needed
Defer the creation of the connection `Configuration` to right before
when it is used, so that `Credentials` are only ever passed in at that
point in time.
This prevents the `Passphrase` from being turned into a `String` for the
lifetime of the `NetHsm`.
Consequently we are now storing all available `Credentials` in a
`HashMap` and refer to the current `Credentials` by `String` (matching a
key in the `HashMap`).
The required `Agent` and `Url` are now also members of the `NetHsm`
struct and the latter can be set using the `NetHsm::set_url` function.

As a drive-by change the `NetHsm::set_url` function now takes a
`nethsm::Url` instead of a `String`.

Signed-off-by: David Runge <dvzrv@archlinux.org>