test: Pin nethsm container image version to c16fe4ed
We are experiencing issues with uploading of update files never
returning.
This is due to us reusing nethsm-py-sdk's dummy update file, which is
now too old to be able to update the container.
Changes to the tagging scheme, as well as a central, versioned location
for testing update files have been requested upstream in
https://github.com/Nitrokey/nethsm/issues/8

Fixes https://gitlab.archlinux.org/archlinux/signstar/-/issues/32

Signed-off-by: David Runge <dvzrv@archlinux.org>