docs: Adapt existing documentation for the use of namespaces
Adapt documentation for all `NetHsm` methods to reflect their respective
use by users in namespaces.
Standardize the mentioning of roles to directly refer to `UserRole`
variants, as it provides better insights for users of the library.
Use "NetHSM" instead of "device" throughout documentation, as the
backend might be a container.
Fix various wording issues in many functions.
Extend code examples to reflect the function's use with namespaces and
provide more complete examples, so that users are more easily able to
understand user and key requirements for all actions.
Standardize towards not using inline links as those are terrible for
text wrapping and reuse.

Signed-off-by: David Runge <dvzrv@archlinux.org>