fix!: Rename function for creating an OpenPGP certificate
The function `NetHsm::add_openpgp_cert` only creates an OpenPGP
certificate and does not in fact persist the resulting certificate in
the NetHSM.
Persisting the certificate using `NetHsm::import_key_certificate` would
require switching to a user in the Administrator role and the function
would therefore require using both users in the Operator and in the
Administrator role.
Renaming the function to `NetHsm::create_openpgp_cert` and adapting
its documentation is the more future-proof solution, as it narrows the
permission scope and reflects the current implementation properly.

Signed-off-by: David Runge <dvzrv@archlinux.org>