feat: Allow providing the global `--user` option multiple times
With `nethsm openpgp add` the CLI has gained a subcommand, that requires
both users in the Operator and Administrator role.
When multiple Operator and Administrator users are configured in the
configuration file, the CLI is not able to choose automatically for this
scenario and without being able to provide `--user` multiple times, it
was not possible to provide the matching user(s) either.

Allowing to provide `--user` multiple times enables providing a set of
specific users to use for commands that require more than one user role.

Signed-off-by: David Runge <dvzrv@archlinux.org>