feat: Allow providing global `--passphrase-file` option multiple times
When allowing the `--user` option to be provided multiple times, this
allows for the automation of commands requiring multiple user roles
(e.g. `nethsm openpgp add`), if configuration file credentials also
carry passphrases.
To allow automating the use of these commands on the command-line as
well, we need to be able to provide `--passphrase-file` multiple times,
too.

Signed-off-by: David Runge <dvzrv@archlinux.org>