fix: Use correct function for constructing MPIs
Previous implementation emitted zero-padded MPIs with wrong bit sizes
which would trip other OpenPGP implementations.

This happened only when the resulting signature had a zero byte at the
beginning. It happened rarely and that's why the test was flaky.

The root cause was the use of an `pgp::types::Mpi::from_slice`
function which produces invalid MPIs so the code was refactored to use
a single safe function and the call has been forbidden by a clippy rule.

See: https://github.com/rpgp/rpgp/issues/401
Fixes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/67
Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>