feat: Add just recipes for building and running OS images
Using the `create-image-signing-key` recipe it is possible to create a
custom signing key (used for verity and secure boot).

The `build-image` recipe is used to build the "SignstarOS" images using
mkosi and signs them using a predefined OpenPGP key.
The certificate for the OpenPGP key is currently used to provide a valid
keyring file (used for automatic updates using systemd-sysupdate).

With the `run-image` the currently built "SignstarOS" image is run using
QEMU (for test purposes).

Add all required packages for building and running the "SignstarOS"
images using mkosi.

Signed-off-by: David Runge <dvzrv@archlinux.org>