feat: Add build-time configuration tool for signstar host
Add executable `signstar-configure-build` to create system users and
their integration for signstar hosts during build-time.
Users are created without a passphrase with the help of `useradd` and
unlocked using `usermod`.
User home directories are created in a dedicated directory with the help
of `tmpfiles.d`.
Afterwards, if available for the specific user mapping, SSH
configuration is created based on system-wide drop-ins, that define
which SSH keys are authorized for authentication and sets a command,
which is enforced upon login.

Fixes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/78
Signed-off-by: David Runge <dvzrv@archlinux.org>