test(mkosi): Add profile for testing of latest local executables
As we want to be able to test executables such as
`signstar-configure-build` in CI on each change, we need a separate
mkosi profile to be able to provide the latest executable and an example
configuration without interfering with the stable image build.

The `local-testing` mkosi profile allows copying more files into the
image which will shadow whatever is installed into the image.

Signed-off-by: David Runge <dvzrv@archlinux.org>