feat: Render top-level docs using mdbook
Rely on mdbook to render top-level documentation in a concise manner.
Using the `build-book` just recipe the mdbook (including adhoc addition
of js requirements for nermaid graphs) is built from a preselected list
of documentation files.
Afterwards all Rust documentation is collected in a `rustdoc`
subdirectory of the shared documentation output directory.

The `watch-book` and `serve-book` just recipes enable a worklow in which
relevant files are watched and the output dir is continuously served
using an adhoc webserver.

Fixes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/124
Signed-off-by: David Runge <dvzrv@archlinux.org>