feat: Rely on `aws-lc-rs` as default crypto provider for `rustls`
The `ring` feature for the `rustls` crate is no longer feasible now that
the upstream project is unmaintained.
Switch the direct use of `rustls` to rely on the default features, which
include the `aws-lc-rs` crypto provider (an AWS maintained fork of
OpenSSL).
Allow the OpenSSL license to be able to use the `aws-lc-rs` feature.

Fixes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/152
Signed-off-by: David Runge <dvzrv@archlinux.org>