fix(deps)!: upgrade rpgp to 0.16.0
rPGP 0.16 uses more RustCrypto-native types in the API avoiding exposure of raw MPIs.
Due to RustCrypto types being used the test suite has to be adjusted to compare public key bytes instead of their MPI encodings.

Sadly, due to the incompatibility of our `digest` crate version (0.11) and theirs (0.10) a newtype wrapper had to be introduced.
This should be only a temporary measure until a new stable release of `digest` is released and rPGP updates to that.

Additionally this MR makes all functions in the `openpgp` module return `openpgp::Error` (instead of the `crate::Error`).
There are also smaller refactorings such as avoiding panics in unreachable branches (rewritten to better match the values) and additional logging in failed conditions.

Closes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/63
Closes: https://gitlab.archlinux.org/archlinux/signstar/-/merge_requests/247
Co-authored-by: Heiko Schaefer <heiko@schaefer.name>
Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>