feat!: Allow detecting when the certificate has not been uploaded
Previously, checking for the certificate would trigger an error and it would not be obvious if that was a communication failure or just that the certificate has not been uploaded.

This MR adjusts the code and all clients to use `None` when the certificate is missing and `Some(Vec<u8>)` for valid, present certificate bytes.

Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>