test: Collect coverage produced by tests run (partially) as other user
Some of the containerized integration tests run as a different user.
The test is run as root in a containerized environment and a specific
subtask is run e.g. by an unprivileged user.

This change ensures that if `.profraw` files are created by running
payloads as an unprivileged user, those coverage files are saved to
a directory exposed by the `CARGO_LLVM_COV_TARGET_DIR` environment
variable.

Signed-off-by: David Runge <dvzrv@archlinux.org>