feat!: Support using different HSM backends in `SignstarConfig`
Add the enum `BackendConnection` and use it instead of
`nethsm::Connection` in `SignstarConfig`.
This is a breaking change for the configuration file format.

Whereas previously a NetHSM connection was configured with

```toml
[[connections]]
tls_security = "Unsafe"
url = "https://127.0.0.1:8080"
```

it is now configured with

```toml
[[connections]]
nethsm = { tls_security = "Unsafe", url = "https://localhost:8443/api/v1/" }
```

Signed-off-by: David Runge <dvzrv@archlinux.org>