refactor!: More generically represent config and backend state
Introduces the `StateHandling` trait, which allows for the infallible
comparison of state from differing origins.
For this, the new `StateComparisonReport` enum is used, which may
indicate incompatibility, failure or success.

The combined `State` struct for NetHSM backends is replaced by two
`StateHandling` implementations (`SignstarConfigNetHsmState` and
`NetHsmState`) to more clearly distinguish configuration and backend.

Fixes: https://gitlab.archlinux.org/archlinux/signstar/-/issues/263
Signed-off-by: David Runge <dvzrv@archlinux.org>